+++ packages/logger/package.json
@@ -11,8 +11,8 @@
   "module": "./dist/index.mjs",
   "exports": {
     ".": "./dist/index.mjs",
-    "./providers/ConsoleProvider": "./dist/Providers/ConsoleProvider.mjs",
-    "./providers/JsonProvider": "./dist/Providers/JsonProvider.mjs",
+    "./drivers/ConsoleProvider": "./dist/Drivers/ConsoleProvider.mjs",
+    "./drivers/JsonProvider": "./dist/Drivers/JsonProvider.mjs",
     "./package.json": "./package.json"
   },
   "scripts": {
