+++ packages/logger/test/BaseLogger.test.ts
@@ -1,8 +1,8 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { Container } from '@vercube/di';
 import { BaseLogger, Logger, type LoggerTypes } from '../src';
-import { JSONProvider } from '../src/Providers/JsonProvider';
-import { ConsoleProvider } from '../src/Providers/ConsoleProvider';
+import { JSONProvider } from '../src/Drivers/JsonProvider';
+import { ConsoleProvider } from '../src/Drivers/ConsoleProvider';
 
 describe('BaseLogger', () => {
   let logger: Logger;
